# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ParkControl Backend API",
    "description": "API Gateway que consume la API Central de estacionamientos",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://parkingrepositorio-backend.onrender.com",
      "description": "Produccion (Render)"
    }
  ],
  "paths": {
...
    "/api/estacionamientos/{id}": {
      "get": { "tags": ["Estacionamientos"], "summary": "Obtener por ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Estacionamiento encontrado" } } },
      "put": { "tags": ["Estacionamientos"], "summary": "Actualizar", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "codigoPlaza": { "type": "string" }, "tipo": { "type": "string" }, "idApartamento": { "type": "integer" } } } } } }, "responses": { "200": { "description": "Actualizado" } } },
      "delete": { "tags": ["Estacionamientos"], "summary": "Eliminar", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "204": { "description": "Eliminado" } } }
    },
    "/api/zonas-estacionamiento": {
      "get": { "tags": ["Zonas Estacionamiento"], "summary": "Listar todas", "responses": { "200": { "description": "Lista de zonas" } } },
      "post": {
        "tags": ["Zonas Estacionamiento"], "summary": "Crear",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "codigo": { "type": "string" },
                  "nombre": { "type": "string" },
                  "idCondominio": { "type": "integer" }
                },
                "required": ["codigo", "idCondominio"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Creado" } }
      }
    },
    "/api/zonas-estacionamiento/{id}": {
      "get": { "tags": ["Zonas Estacionamiento"], "summary": "Obtener por ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Zona encontrada" } } },
      "put": { "tags": ["Zonas Estacionamiento"], "summary": "Actualizar", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "codigo": { "type": "string" }, "nombre": { "type": "string" }, "idCondominio": { "type": "integer" } } } } } }, "responses": { "200": { "description": "Actualizado" } } },
      "delete": { "tags": ["Zonas Estacionamiento"], "summary": "Eliminar", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "204": { "description": "Eliminado" } } }
    },
    "/api/detalles-plaza": {
      "get": { "tags": ["Detalles Plaza"], "summary": "Listar todos", "responses": { "200": { "description": "Lista de detalles" } } },
      "post": {
        "tags": ["Detalles Plaza"], "summary": "Crear",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idEstacionamiento": { "type": "integer" },
                  "idZona": { "type": "integer" },
                  "estadoOcupacion": { "type": "string" },
                  "idVehiculoActual": { "type": "integer" }
                },
                "required": ["idEstacionamiento", "estadoOcupacion"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Creado" } }
      }
    },
    "/api/detalles-plaza/{id}": {
      "get": { "tags": ["Detalles Plaza"], "summary": "Obtener por ID", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Detalle encontrado" } } },
      "put": { "tags": ["Detalles Plaza"], "summary": "Actualizar", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "idEstacionamiento": { "type": "integer" }, "idZona": { "type": "integer" }, "estadoOcupacion": { "type": "string" }, "idVehiculoActual": { "type": "integer" } } } } } }, "responses": { "200": { "description": "Actualizado" } } },
      "delete": { "tags": ["Detalles Plaza"], "summary": "Eliminar", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "204": { "description": "Eliminado" } } }
    },
...
  }
}